# zina_app/config/openapi.yaml
openapi: 3.0.2
info:
  title: ZINA API
  description: ZINA API documentation (Public + Admin)
  version: 2.0.0
servers:
  - url: /
paths:
  /api/menu:
    get:
      tags: [Public, Menu]
      summary: Get full menu organized by category
      responses:
        '200':
          description: Menu organized by category key
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    $ref: '#/components/schemas/MenuProduct'
        '500':
          $ref: '#/components/responses/ServerError'

  /api/categories:
    get:
      tags: [Public, Categories]
      summary: Get all categories
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
        '500':
          $ref: '#/components/responses/ServerError'

  /api/products/{product_id}:
    get:
      tags: [Public, Products]
      summary: Get product by id
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError'

  /api/formulas:
    get:
      tags: [Public, Formulas]
      summary: Get formulas discounts configuration
      responses:
        '200':
          description: Formulas discounts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormulasConfig'
        '500':
          $ref: '#/components/responses/ServerError'

  /api/order:
    post:
      tags: [Public, Orders, Ordering]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '200':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderCreatedResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError'

  /api/admin/settings:
    get:
      tags: [Admin, Settings]
      summary: Get site settings
      responses:
        '200':
          description: Settings payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSettings'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags: [Admin, Settings]
      summary: Update site settings
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '500':
          $ref: '#/components/responses/ServerError'

  /api/admin/formulas:
    get:
      tags: [Admin, Formulas]
      summary: Get formulas discounts configuration (admin)
      responses:
        '200':
          description: Formulas discounts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormulasConfig'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags: [Admin, Formulas]
      summary: Update formulas discounts configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FormulasConfig'
      responses:
        '200':
          description: Updated config
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/FormulasConfig'
        '500':
          $ref: '#/components/responses/ServerError'

components:
  responses:
    ServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Something went wrong
        message:
          type: string
          example: Something went wrong

    StatusResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        message:
          type: string
          example: Settings updated

    FormulasConfig:
      type: object
      properties:
        discounts:
          type: object
          properties:
            EP:
              type: integer
              example: 200
            PD:
              type: integer
              example: 100
            EPD:
              type: integer
              example: 300
            EPDB:
              type: integer
              example: 500

    MenuProduct:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: number
          format: float
        description:
          type: string
        image:
          type: string
        prep_time:
          type: string
        category:
          type: string
        options:
          type: array
          items:
            $ref: '#/components/schemas/ProductOption'

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        image:
          type: string
          nullable: true
        products_count:
          type: integer

    ProductOption:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: number
          format: float

    Product:
      type: object
      properties:
        id:
          type: integer
        category_id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        image_url:
          type: string
          nullable: true
        is_available:
          type: boolean
        category_name:
          type: string
          nullable: true
        options:
          type: array
          items:
            $ref: '#/components/schemas/ProductOption'

    OrderItemRequest:
      type: object
      required: [product_id, quantity]
      properties:
        product_id:
          type: integer
          example: 12
        quantity:
          type: integer
          example: 2
        option_ids:
          type: array
          items:
            type: integer
          default: []

    CreateOrderRequest:
      type: object
      required: [items]
      properties:
        user_id:
          type: string
          nullable: true
          description: UUID user id (optional for guest)
          example: 2f6d0b2f-7d2e-4b7b-9f11-6d2f9f1a3a11
        pickup_time:
          type: string
          format: date-time
          nullable: true
          example: '2026-03-19T18:30:00'
        prep_time_minutes:
          type: integer
          default: 15
        discount_amount:
          type: integer
          default: 0
          description: Discount in FCFA applied to order total
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemRequest'

    OrderCreatedResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        order_id:
          type: integer
          example: 123
        user_id:
          type: string
          nullable: true
        total_amount:
          type: number
          format: float
        order_status:
          type: string
        pickup_time:
          type: string
          nullable: true
        prep_time_minutes:
          type: integer
        items:
          type: array
          items:
            type: object
            properties:
              product_id: { type: integer }
              product_name: { type: string }
              quantity: { type: integer }
              unit_price: { type: number, format: float }
              options:
                type: array
                items:
                  type: object
                  properties:
                    option_id: { type: integer }
                    option_name: { type: string }
                    additional_price: { type: number, format: float }
        message:
          type: string

    BusinessHours:
      type: object
      additionalProperties: true

    AdminSettings:
      type: object
      properties:
        company_name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        hours:
          $ref: '#/components/schemas/BusinessHours'
        service_fee:
          type: number
          format: float
        tax_rate:
          type: number
          format: float
        payment_methods:
          type: array
          items:
            type: string